Hello World in assembler on the HW

+++ src/test/scala/wildcat/WildcatTest.scala
@@ -21,10 +21,14 @@ class WildcatTest() extends AnyFlatSpec with ChiselScalatestTester {
   progs.foreach(p => {
     println(s"Running test $p")
     s"Program $p" should "pass" in {
-      "rm a.out".!
-      s"make app APP=$p".!
+      var app = p
+      if (p.endsWith(".s")) {
+        "rm a.out".!
+        s"make app APP=$p".!
+        app = "a.out"
+      }
       //    test(new FiveCats(Array("a.out"))).withAnnotations(Seq(WriteVcdAnnotation)) {
-      test(new WildcatTestTop("a.out")).withAnnotations(Seq(WriteVcdAnnotation)) {
+      test(new WildcatTestTop(app)).withAnnotations(Seq(WriteVcdAnnotation)) {
         d => {
           var stop = false
           var cnt = 0
